requirements.txt
streamlit
yt-dlp
faster-whisper
python-docx
ffmpeg-python
pydantic
watchdog
python-dotenv
openai-whisper
torch
numpy
